Refactor UI and data handling in various components
- Updated `AppShell` to improve flyout item management.
- Modified `EncounterDB` methods for better data handling.
- Adjusted `SearchSettings` for proper byte casting.
- Updated `PKHeX.MAUI.csproj` to reference a newer `PKHeX.Core`.
- Changed layouts in `ALMSettings.xaml` and `LegalitySettings.xaml` to use `ScrollView`.
- Enhanced UI in `ALMSettings` and `LegalitySettings` with `propertyGrid`.
- Improved organization in `PKHeXSettings.xaml` with a new namespace.
- Updated `PSettings` to use a `StartPage` enum for better type safety.
- Refactored `propertyGrid` for correct property display.
- Enhanced `BatchEditor` for better property type handling.
- Updated `MainPage` navigation to utilize the new `StartPage` enum.
- Removed `PrioritizeGame` and `PrioritizeGameVersion` from `APILegality`.

+++ EncounterDB/SearchSettings.xaml.cs
@@ -61,7 +61,7 @@ public partial class SearchSettings : ContentPage
             Version = (GameVersion)((ComboItem?)EncVersion.SelectedItem??Any).Value,
             Nature = (EncounterSettings.UsePkEditorAsCriteria ? pk.Nature : 0),
             Ability = (EncounterSettings.UsePkEditorAsCriteria ? pk.Ability : 0),
-            Level = (EncounterSettings.UsePkEditorAsCriteria ? pk.CurrentLevel : 0),
+            Level = (EncounterSettings.UsePkEditorAsCriteria ? pk.CurrentLevel : (byte)0),
             Item = (EncounterSettings.UsePkEditorAsCriteria ? pk.HeldItem : 0)
         };
         encSettings.AddMove((ushort)((ComboItem?)EncMove1.SelectedItem??Any).Value);
